Parallax now moves when the player moves. Not automatically.

+++ Platformer 2D/Motor2D/j1Map.h
@@ -57,7 +57,7 @@ struct ImageLayer
 	int				width;
 	int				height;
 	fPoint			position;
-	float			speed = 0.0f;
+	int				speed = 0;
 
 };
